+++ src/components/solar.js
@@ -1,5 +1,4 @@
-import React, { useState } from "react";
-
+import React from "react";
 import {
   Typography,
   Toolbar,
@@ -11,7 +10,6 @@ import {
   Accordion,
   AccordionSummary,
   AccordionDetails,
-  Button,
 } from "@mui/material";
 import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
